Ops: Setup Docker and Docker Compose
Created a single Dockerfile that compiles both the
backend and frontend for the project. After compilation
the frontend is configured to be served statically from the backend.

Changed frontend bundler to Vite, to ease serving from backend.

Configured docker compose with our web app and a postgress database.
For now the database uses a .env file for the password.