Code: Tools for password hashing
Added a tools package for miscelaneous things we may need.

Wrote functions for generating and testing the hashes of user passwords.
We use argon2 on the inside but the algorithm is invisible to the end
user, outside this package it's enough to call GeneratePasswordHash when
a user is created and doesPasswordMatch to test on login.