Feat: Preliminary user registration and login
Start of registration and logic setup.

Created a base_user model and controller to handle the very basic
functionality for both trainers and users.

Added a session manager to keep track of the logged in users and
possibly of other "per session" data.

Moved the database into it's own file, exposing the orm through Orm().
Also moved the database credentials to the compose and .env files.

Fixed a parameter in password, greatly reducing the generated password
size.