Feat: Protected routes and login using cookies
Made a protectedRoute component, which can be used to reroute a user if
they haven't logged in yet.

Added ContinueSession to backend. Now if a user has a valid, living
session cookie we can use it to log them in without asking for their
credentials again. For now this endpoint is only accessed on the /login
view but we could probably make it available on other places.