Fix: ProtectedRouteComponent now attempts autologin
If a user has a session cookie and the Redux state marks them
as not logged in then we try to login with the cookie instead of
instantly redirecting them to the login page.

Previously accessing a protected url like "editTrainerProfile" would
first redirect the user to login and then attempt the autologin, if the
session existed then the user would be confusingly redirected to
"messages".

Removed React Strict mode since it was launching multiple login attempts
(cause it renders everything twice for some reason).

Moved the toast container to main, now displaying an error can be done
by just calling toast()