Ops: Now calls to /api from frontend are proxied to backend
Configured Vite to proxy petitions to backend, removing the need for
CORS configuration and also preparing us for deployment, since hardcoded
localhost urls are no longer needed.

For now on instead of calling a route on backend as
'http://localhost:9090/something' we use '/api/something'

Note that the websocket route used in chat needed a custom proxy rule.