Code: Trainers and athletes are now separated
Completed the separation between trainers and athletes.
Now trainers have /trainerProfile and athletes /userProfile as their
respective homepages, and their views may be customized further.

Now the register views create not only the baseUser but also the trainer
or athlete entity on the db.

Login api endpoints now return a _kind_ attribute with the type of user
being logged in.

Fixed the User and Trainer models, now both respect their belongsTo
relationship with BaseUser.

Protected route can now restrict a kind of user from seeing a view, to
do so we use the property _kindsAllowed_. The default is to allow all
kinds of users as long as they are logged in.

Finally, added the kind attribute to the userSlice in redux. We can now
get the kind of user we are dealing with directly from the front-end.