Code: Athletes can see their trainer requests
After an athlete requests a trainer the request and it's status
(waiting, denied, accepted) is shown in the /requestTrainer view.

Changed the request model to consider all posible responses from a
trainer, denying or accepting that is. These are stored as strings due
to the lack of support of enums by our ORM.

Created a controller for requests and built an endpoint to display the
requests for an athlete.

Made a new view for requesting trainers and moved the trainer search
component to it, alongside the ShowRequests component.

Finally, added a green to our color palette since we need it to display
success in some parts of the page.