Code: Trainer specialty editing completed
A trainer now may change their specialties in editTrainerProfile, by
clicking buttons with the name of the tags.

Made a controller for the specialties & routes for it.

Removed the preliminary view we had for tag editing.

Extracted the base user editing functionality into it's own component
and put the tag editing component alongside it in the editTrainerProfile
view.