Code: Trainers can now like posts
Added a new m-m relationship between users and posts to act as likes for
the post.

Made endpoints to see if a user has liked a given post and for that same
user to like a post, together they make a basic like button.

Added like buttons to the post views.